feat: move to compute namespace and add undelegate cmd
