+++ packages/cli/src/commands/compute/app/stop.ts
@@ -1,31 +1,25 @@
-import { Command, Args, Flags } from "@oclif/core";
-import { createAppClient } from "../../client";
-import { commonFlags } from "../../flags";
+import { Command, Args } from "@oclif/core";
+import { createAppClient } from "../../../client";
+import { commonFlags } from "../../../flags";
 import { getEnvironmentConfig, getOrPromptAppID } from "@layr-labs/ecloud-sdk";
 import chalk from "chalk";
 
-export default class AppLifecycleTerminate extends Command {
-  static description =
-    "Terminate app (terminate GCP instance) permanently";
+export default class AppLifecycleStop extends Command {
+  static description = "Stop running app (stop GCP instance)";
 
   static args = {
     "app-id": Args.string({
-      description: "App ID or name to terminate",
+      description: "App ID or name to stop",
       required: false,
     }),
   };
 
   static flags = {
     ...commonFlags,
-    force: Flags.boolean({
-      required: false,
-      description: "Force termination without confirmation",
-      default: false,
-    }),
   };
 
   async run() {
-    const { args, flags } = await this.parse(AppLifecycleTerminate);
+    const { args, flags } = await this.parse(AppLifecycleStop);
     const app = await createAppClient(flags);
 
     // Get environment config
@@ -42,18 +36,16 @@ export default class AppLifecycleTerminate extends Command {
         environment: flags["environment"]!,
         privateKey: flags["private-key"],
         rpcUrl,
-        action: "terminate",
+        action: "stop",
       }
     );
 
-    const res = await app.terminate(appId, {
-      force: flags.force,
-    });
+    const res = await app.stop(appId);
 
     if (!res.tx) {
-      this.log(`\n${chalk.gray(`Termination aborted`)}`);
+      this.log(`\n${chalk.gray(`Stop aborted`)}`);
     } else {
-      this.log(`\n✅ ${chalk.green(`App terminated successfully`)}`);
+      this.log(`\n✅ ${chalk.green(`App stopped successfully`)}`);
     }
   }
 }
