+++ packages/cli/src/commands/compute/app/terminate.ts
@@ -1,25 +1,31 @@
-import { Command, Args } from "@oclif/core";
-import { createAppClient } from "../../client";
-import { commonFlags } from "../../flags";
+import { Command, Args, Flags } from "@oclif/core";
+import { createAppClient } from "../../../client";
+import { commonFlags } from "../../../flags";
 import { getEnvironmentConfig, getOrPromptAppID } from "@layr-labs/ecloud-sdk";
 import chalk from "chalk";
 
-export default class AppLifecycleStop extends Command {
-  static description = "Stop running app (stop GCP instance)";
+export default class AppLifecycleTerminate extends Command {
+  static description =
+    "Terminate app (terminate GCP instance) permanently";
 
   static args = {
     "app-id": Args.string({
-      description: "App ID or name to stop",
+      description: "App ID or name to terminate",
       required: false,
     }),
   };
 
   static flags = {
     ...commonFlags,
+    force: Flags.boolean({
+      required: false,
+      description: "Force termination without confirmation",
+      default: false,
+    }),
   };
 
   async run() {
-    const { args, flags } = await this.parse(AppLifecycleStop);
+    const { args, flags } = await this.parse(AppLifecycleTerminate);
     const app = await createAppClient(flags);
 
     // Get environment config
@@ -36,16 +42,18 @@ export default class AppLifecycleStop extends Command {
         environment: flags["environment"]!,
         privateKey: flags["private-key"],
         rpcUrl,
-        action: "stop",
+        action: "terminate",
       }
     );
 
-    const res = await app.stop(appId);
+    const res = await app.terminate(appId, {
+      force: flags.force,
+    });
 
     if (!res.tx) {
-      this.log(`\n${chalk.gray(`Stop aborted`)}`);
+      this.log(`\n${chalk.gray(`Termination aborted`)}`);
     } else {
-      this.log(`\n✅ ${chalk.green(`App stopped successfully`)}`);
+      this.log(`\n✅ ${chalk.green(`App terminated successfully`)}`);
     }
   }
 }
